Move unix socket support to io module
